Added smoke detection service
Smoke detection service is used to detect smoke in the given video. So if there is any smoke on any given floor or corridor it will help providing an alert to all the members of the society that a fire or a cause of fire has occurred. So they can take the respected precautions